requirements.txt
albumentations==0.4.6
numpy
pillow
tqdm
